feat: pass isTypeScript and isDeno to remix init
this is ran after create converts the project to js if they requested
