+++ packages/remix-dev/cli/commands.ts
@@ -48,57 +48,23 @@ export async function create({
     useTypeScript,
     githubToken,
   });
-
-  let initScriptDir = path.join(projectDir, "remix.init");
-  let hasInitScript = await fse.pathExists(initScriptDir);
-
   spinner.stop();
   spinner.clear();
-
-  if (hasInitScript) {
-    if (installDeps) {
-      console.log("💿 Running remix.init script");
-      await init(projectDir);
-      await fse.remove(initScriptDir);
-    } else {
-      console.log();
-      console.log(
-        colors.warning(
-          "💿 You've opted out of installing dependencies so we won't run the " +
-            "remix.init/index.js script for you just yet. Once you've installed " +
-            "dependencies, you can run it manually with `npx remix init`"
-        )
-      );
-      console.log();
-    }
-  }
-
-  let relProjectDir = path.relative(process.cwd(), projectDir);
-  let projectDirIsCurrentDir = relProjectDir === "";
-
-  if (projectDirIsCurrentDir) {
-    console.log(
-      `💿 That's it! Check the README for development and deploy instructions!`
-    );
-  } else {
-    console.log(
-      "💿 That's it! `cd` into " +
-        colors.logoGreen(path.resolve(process.cwd(), projectDir)) +
-        " and check the README for development and deploy instructions!"
-    );
-  }
 }
 
 export async function init(projectDir: string) {
   let initScriptDir = path.join(projectDir, "remix.init");
   let initScript = path.resolve(initScriptDir, "index.js");
 
+  let isTypeScript = fse.existsSync(path.join(projectDir, "tsconfig.json"));
+  let isDeno = fse.existsSync(path.join(projectDir, "deno.json"));
+
   if (await fse.pathExists(initScript)) {
     // TODO: check for npm/yarn/pnpm
     execSync("npm install", { stdio: "ignore", cwd: initScriptDir });
     let initFn = require(initScript);
     try {
-      await initFn({ rootDirectory: projectDir });
+      await initFn({ rootDirectory: projectDir, isTypeScript, isDeno });
     } catch (error) {
       if (error instanceof Error) {
         error.message = `${colors.error("🚨 Oops, remix.init failed")}\n\n${
